+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to the "neo4j-vscode" extension will be documented in this file.
 
+## [0.1.9]
+
+- Added support for Cypher code blocks in Markdown files
+- Automatically selects the text within the code block when running a query if no selection is made
+
+```cypher
+match (n) return n limit 1
+```
+
+[Reference](https://stackoverflow.com/a/76239666/3876654)
+
 ## [0.1.8]
 
 - Applied case insensitivity to parameter commands